find pwd for corp
