+++ public/static/lang/en-us.js
@@ -141,6 +141,7 @@ module.exports = {
         noSession: 'System error, please refresh the email domain list and try again',
         sameEmailSuffix: 'Email domain already exists, please do not add it repeatedly',
         no_link_or_unsigned: 'Invalid binding information',
+        unknown_email: 'The email is not an manager',
     },
     header: {
         home: 'Home',
@@ -187,9 +188,8 @@ module.exports = {
         newPwd_length_err: 'the length of password should be between 6 and 16',
         newPwd_contains_Illegal_character: 'The new password contains illegal characters',
         email_input_holder: 'Please input email',
-        find_pwd_step1: 'Verify email',
-        find_pwd_step2: 'Reset password',
-        find_pwd_step3: 'Done',
+        find_pwd_step1: 'Reset password',
+        find_pwd_step2: 'Done',
         resetSuccess: 'Reset successfully',
         goToLogin: 'Go to login',
         managerName: 'Manager Name',
