+++ public/static/lang/zh-cn.js
@@ -141,6 +141,7 @@ module.exports = {
         noSession: '系统错误，请刷新邮箱域名列表后再试',
         sameEmailSuffix: '邮箱域名已存在，请勿重复添加',
         no_link_or_unsigned: '无效的绑定信息',
+        unknown_email: '该邮箱不是管理员',
     },
     header: {
         home: '首页',
@@ -187,9 +188,8 @@ module.exports = {
         newPwd_length_err: '新密码长度必须是6-16位',
         newPwd_contains_Illegal_character: '新密码包含非法字符',
         email_input_holder: '请输入邮箱',
-        find_pwd_step1: '验证邮箱',
-        find_pwd_step2: '重置密码',
-        find_pwd_step3: '完成',
+        find_pwd_step1: '重置密码',
+        find_pwd_step2: '完成',
         resetSuccess: '重置成功',
         goToLogin: '去登陆',
         managerName: '管理员姓名',
