+++ src/components/FindPwdSteps.vue
@@ -3,7 +3,6 @@
         <el-steps :active="steping" align-center>
             <el-step :title="$t('corp.find_pwd_step1')" description=""></el-step>
             <el-step :title="$t('corp.find_pwd_step2')" description=""></el-step>
-            <el-step :title="$t('corp.find_pwd_step3')" description=""></el-step>
         </el-steps>
     </div>
 </template>
