+++ src/util/api.js
@@ -32,8 +32,9 @@ export const getPrivacy = '/v1/getPrivacy';                                     
 export const corporationPdf = '/v1/corporation-pdf';                                //上传pdf签名页        方式patch  参数：/link_id/email
 export const resend_pdf = '/v1/corporation-signing';                                //重发企业签名pdf      方式post  参数：/link_id/email
 export const addCla = 'v1/cla';                                                     //增加其他语言的CLA     方式post  参数：/:link_id/:apply_to
-export const findPwdSendCode = 'v1/findPwdSendCode';                                //忘记密码发送验证码     方式  参数：
 export const addSubEmail = 'v1/corporation-email-domain';                           //添加邮箱后缀       方式post   参数：verification_code,sub_email
+export const findPwdSendCode = 'v1/verification-code';                              //忘记密码发送验证码     方式post  参数：/:platform/:org_id:repo_id/:email
+export const submitVerifyCode = '/v1/corporation-manager/retrieve';                 //忘记密码提交新密码     方式 post  参数：{code: ,email: ,org_repo: {platform: ,org_id: ,repo_id: },password: }
 export const submitNewPassword = 'v1/submitNewPassword';                            //提交新密码        方式  参数：
 export const getBlankSignature = 'v1/org-signature/blank';
 export const getCLAPdf = 'v1/cla';                                                  //下载cla的PDF文件    方式 get  参数   /:link_id/:apply_to/:language/:hash
