+++ src/views/CorporationManagerLogin.vue
@@ -5,11 +5,11 @@
                 <div class="formBack">
                     <el-form :model="ruleForm" status-icon :rules="rules" ref="ruleForm" label-width="0">
                         <el-form-item :required="true" prop="userName">
-                            <el-input v-model="ruleForm.userName" autocomplete="off" :placeholder="$t('corp.id')"
+                            <el-input clearable="" v-model="ruleForm.userName" autocomplete="off" :placeholder="$t('corp.id')"
                                       @keydown.native="pressEnter"></el-input>
                         </el-form-item>
                         <el-form-item :required="true" label="" prop="pwd">
-                            <el-input type="password" v-model="ruleForm.pwd" autocomplete="off"
+                            <el-input clearable="" type="password" v-model="ruleForm.pwd" autocomplete="off"
                                       :placeholder="$t('corp.pwd')" @keydown.native="pressEnter"></el-input>
                         </el-form-item>
                         <el-form-item style="text-align: right">
@@ -93,7 +93,7 @@
                 }
             },
             findPwd() {
-                // this.$router.push('/password')
+                this.$router.push('/password')
             },
             login(userName, pwd) {
                 let obj = {
