+++ src/views/LinkedRepo.vue
@@ -200,7 +200,6 @@
                 orgTableData: [],
                 boundTableData: [],
                 url: '',
-                signRouter: this.$store.state.signRouter,
                 pdfSrc: '',
                 numPages: undefined,
                 docInfo: {},
@@ -433,7 +432,7 @@
                 }
                 let base64Params = util.strToBase64(params)
                 let address = window.location.href.split('/linkedRepo')[0];
-                let url = `${address}${this.signRouter}/${base64Params}`
+                let url = `${address}${SIGN_ROUTER}/${base64Params}`
                 let copyInput = document.createElement("input");
                 copyInput.value = url;
                 document.body.appendChild(copyInput);
@@ -451,7 +450,7 @@
                     params = `${row.platform.toLowerCase()}/${row.org_id}`
                 }
                 let base64Params = util.strToBase64(params)
-                let url = `${this.address}${this.signRouter}/${base64Params}`
+                let url = `${this.address}${SIGN_ROUTER}/${base64Params}`
                 window.open(url)
             },
             submitUpload() {
