+++ src/views/LoginTypeSelect.vue
@@ -15,16 +15,6 @@
                             </el-tooltip>
 
                         </div>
-                        <div>
-                            <el-tooltip :content="$t('index.corp_bt_tip')" placement="bottom" effect="light"
-                                        popper-class="my_tooltip">
-                                <button class="button" @click="submit('corporationManager')">
-                                    {{$t('index.corp_login')}}
-                                </button>
-                            </el-tooltip>
-                        </div>
-
-
                     </div>
                 </el-col>
             </el-row>
@@ -49,10 +39,6 @@
                 this.setLoginTypeAct(loginType);
                 if (loginType === 'orgManager') {
                     this.$router.push('/platformSelect')
-                } else if (loginType === 'corporationManager') {
-                    this.$router.push('/corporationManagerLogin')
-                } else {
-                    this.$router.push('/signCla')
                 }
             },
             clearSessionStorage() {
